services/llm/requirements.txt
python-dotenv==1.0.1
confluent-kafka==2.3.0
pydantic==2.6.4
pydantic-settings==2.2.1
transformers==4.41.2
sentence-transformers==2.5.1
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
coremltools==7.2
google-generativeai==0.4.1
accelerate==0.30.1
nest-asyncio==1.6.0
pip-system-certs==4.0
numpy==1.24.1

tenacity==8.2.2
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-exporter-jaeger==1.21.0
